pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gg-eff-agent-cataloger"
version = "0.1.0"
description = "Production-grade cataloger for AI agent repositories."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyGithub>=2.4.0",
  "GitPython>=3.1.43",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "rich>=13.7.1",
  "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-mock>=3.14.0",
]

[project.scripts]
gg_eff_agent_cataloger = "gg_eff_agent_cataloger.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
